Extracted heap into its own module
